@@ -8,72 +8,73 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using CloudRepublic.BenchMark.Application.Statics;
 
-namespace CloudRepublic.BenchMark.Orchestrator.Services
+namespace CloudRepublic.BenchMark.Orchestrator.Services;
+
+public class BenchMarkTypeService : IBenchMarkTypeService
 {
-    public class BenchMarkTypeService : IBenchMarkTypeService
+    private readonly IBenchMarkService _benchMarkService;
+    private readonly IBenchMarkResultRepository _benchMarkResultRepository;
+
+    public BenchMarkTypeService(IBenchMarkService benchMarkService, IBenchMarkResultRepository benchMarkResultRepository)
     {
-        private readonly IBenchMarkService _benchMarkService;
-        private readonly BenchMarkDbContext _dbContext;
+        _benchMarkService = benchMarkService;
+        _benchMarkResultRepository = benchMarkResultRepository;
+    }
 
-        public BenchMarkTypeService(IBenchMarkService benchMarkService, BenchMarkDbContext dbContext)
+
+    public IEnumerable<BenchMarkType> GetAllTypes()
+    {
+        return BenchMarkTypeGenerator.GetAllTypes();
+    }
+
+    public async Task<List<BenchMarkResult>> RunBenchMarksAsync(IEnumerable<BenchMarkType> benchMarkTypes, int coldCalls = 5, int warmCalls = 10, int delayBetweenCalls = 30)
+    {
+        var results = new List<BenchMarkResult>();
+
+        foreach (var benchMarkType in benchMarkTypes)
         {
-            _benchMarkService = benchMarkService;
-            _dbContext = dbContext;
-        }
+            var tasksCold = new List<Task<BenchMarkResponse>>();
+            var tasksWarm = new List<Task<BenchMarkResponse>>();
 
-
-        public IEnumerable<BenchMarkType> GetAllTypes()
-        {
-            return BenchMarkTypeGenerator.GetAllTypes();
-        }
-
-        public async Task<List<BenchMarkResult>> RunBenchMarksAsync(IEnumerable<BenchMarkType> benchMarkTypes, int coldCalls = 5, int warmCalls = 10, int delayBetweenCalls = 30)
-        {
-            var results = new List<BenchMarkResult>();
-
-            foreach (var benchMarkType in benchMarkTypes)
+            for (int i = 0; i < coldCalls; i++)
             {
-                var tasksCold = new List<Task<BenchMarkResponse>>();
-                var tasksWarm = new List<Task<BenchMarkResponse>>();
-
-                for (int i = 0; i < coldCalls; i++)
-                {
-                    tasksCold.Add(_benchMarkService.RunBenchMarkAsync(benchMarkType.ClientName));
-                }
-
-                await Task.WhenAll(tasksCold);
-
-                await Task.Delay(TimeSpan.FromSeconds(delayBetweenCalls));
-
-                for (int i = 0; i < warmCalls; i++)
-                {
-                    tasksWarm.Add(_benchMarkService.RunBenchMarkAsync(benchMarkType.ClientName));
-                }
-
-                await Task.WhenAll(tasksWarm);
-
-                results.AddRange(ResultConverter.ConvertToResultObject(tasksCold.Select(t => t.Result), benchMarkType, true));
-                results.AddRange(ResultConverter.ConvertToResultObject(tasksWarm.Select(t => t.Result), benchMarkType, false));
-
-            }
-            return results;
-
-        }
-
-        public async Task StoreBenchMarkResultsAsync(IEnumerable<BenchMarkResult> results)
-        {
-            if (!results.Any())
-            {
-                return;
+                tasksCold.Add(_benchMarkService.RunBenchMarkAsync(benchMarkType.ClientName));
             }
 
-            foreach (var result in results)
+            await Task.WhenAll(tasksCold);
+
+            await Task.Delay(TimeSpan.FromSeconds(delayBetweenCalls));
+
+            for (int i = 0; i < warmCalls; i++)
             {
-                _dbContext.BenchMarkResult.Add(result);
+                tasksWarm.Add(_benchMarkService.RunBenchMarkAsync(benchMarkType.ClientName));
             }
 
-            await _dbContext.SaveChangesAsync();
+            await Task.WhenAll(tasksWarm);
+
+            results.AddRange(ResultConverter.ConvertToResultObject(tasksCold.Select(t => t.Result), benchMarkType, true));
+            results.AddRange(ResultConverter.ConvertToResultObject(tasksWarm.Select(t => t.Result), benchMarkType, false));
+
+            for(var i = 0; i < results.Count; i++)
+            {
+                results[i].CallPositionNumber = i;
+            }
+        }
+        return results;
+    }
+
+    public async Task StoreBenchMarkResultsAsync(IEnumerable<BenchMarkResult> results)
+    {
+        if (!results.Any())
+        {
+            return;
+        }
+
+        foreach (var result in results)
+        {
+            await _benchMarkResultRepository.AddBenchMarkResultAsync(result);
         }
     }
 }
